refactor: basic refactoring of the framework for finding obfuscated members
