+++ app/src/main/java/com/gswxxn/camerasnap/dexkit/camera/CameraSnapMembersFinder.kt
@@ -1,0 +1,12 @@
+package com.gswxxn.camerasnap.dexkit.camera
+
+import com.gswxxn.camerasnap.dexkit.base.BaseFinder
+
+/** CameraSnap 类被混淆成员的 Finder **/
+object CameraSnapMembersFinder: BaseFinder(){
+
+    /** 开始查找 **/
+    override fun onFindMembers() {
+
+    }
+}
