+++ app/src/main/java/com/gswxxn/camerasnap/hook/base/BaseHookerWithDexKit.kt
@@ -1,5 +1,6 @@
 package com.gswxxn.camerasnap.hook.base
 
+import com.gswxxn.camerasnap.dexkit.base.BaseFinder
 import com.highcapable.yukihookapi.hook.core.YukiMemberHookCreator
 import com.highcapable.yukihookapi.hook.entity.YukiBaseHooker
 import io.luckypray.dexkit.DexKitBridge
@@ -7,16 +8,11 @@ import io.luckypray.dexkit.builder.MethodInvokingArgs
 import io.luckypray.dexkit.descriptor.member.DexMethodDescriptor
 import java.lang.reflect.Method
 
+/** 使用 DexKit 的 Base Hooker, 在需要的作用域 Hooker 入口继承使用 **/
 abstract class BaseHookerWithDexKit: YukiBaseHooker() {
-    companion object {
-        var isDexKitInit = false
-    }
 
     override fun onHook() {
-        if (!isDexKitInit) {
-            System.loadLibrary("dexkit")
-            isDexKitInit = true
-        }
+        System.loadLibrary("dexkit")
 
         DexKitBridge.create(appInfo.sourceDir)?.use { bridge ->
             onFindMembers(bridge)
@@ -63,5 +59,15 @@ abstract class BaseHookerWithDexKit: YukiBaseHooker() {
         require(flatMap.size == 1) { "uniqueFindMethodInvoking() Error: invokingList must contain exactly one item; Data: $invokingList" }
 
         return flatMap.first().getMethodInstance()
+    }
+
+    /**
+     * 加载 Finder, 将 DexKitBridge 实例传递给 Finder, 并调用 onFindMembers 方法
+     *
+     * @param finder 被操作的 Finder
+     */
+    fun DexKitBridge.loadFinder(finder: BaseFinder) {
+        finder.bridge = this
+        finder.onFindMembers()
     }
 }
