+++ app/src/main/java/com/gswxxn/camerasnap/utils/DexKitHelper.kt
@@ -1,0 +1,19 @@
+package com.gswxxn.camerasnap.utils
+
+/** DexKit 工具类 **/
+object DexKitHelper {
+
+    /** 字段签名 **/
+    object TypeSignature {
+        const val VOID = "V"
+        const val BOOLEAN = "Z"
+        const val BYTE = "B"
+        const val CHAR = "C"
+        const val SHORT = "S"
+        const val INT = "I"
+        const val FLOAT = "F"
+        const val LONG = "J"
+        const val DOUBLE = "D"
+        const val STRING = "Ljava/lang/String;"
+    }
+}
